+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added EasyButton component.
+- Added tracking of map bounds
 
 ## [0.1.4] - 2020-23-09